requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
dnspython==2.4.2
shodan==1.31.0
tweepy==4.14.0
aiohttp==3.13.3
validators==0.22.0
colorama==0.4.6
tabulate==0.9.0
python-dotenv==1.0.0
Pillow==10.3.0
cryptography==42.0.4
stem==1.8.2
PySocks==1.7.1
